new officer weapon

+++ fight.py
@@ -921,7 +921,7 @@ async def fight(uid1, uid2, un1, un2, t, mid):
                 elif int(r.hget(uid2, 'money')) < 200 and checkClan(uid2, building='build3', level=1):
                     m_bonus[0] += 2
                 if weapon2 == 41 and in1 > 0:
-                    m_bonus = [int(m_bonus[0] * 1.34)]
+                    m_bonus = [m_bonus[0] * 1.34]
             elif c2 in (34, 35, 36):
                 m_bonus = [int(m_bonus[0] * (1 + 0.1 * l2))]
             if m_bonus[0] > 0 and can_earn2:
@@ -929,10 +929,10 @@ async def fight(uid1, uid2, un1, un2, t, mid):
                     if choices([1, 0], weights=[s1 / (s1 + s2), 1 - s1 / (s1 + s2)]) == [1]:
                         m_bonus = [m_bonus[0] * 2]
                 if steal == 0:
-                    r.hincrby(uid2, 'money', m_bonus[0])
+                    r.hincrby(uid2, 'money', int(m_bonus[0]))
                     grn = '\n\U0001F4B5 +' + str(m_bonus[0])
                 else:
-                    r.hincrby(uid1, 'money', m_bonus[0])
+                    r.hincrby(uid1, 'money', int(m_bonus[0]))
                     grn = '\n\U0001F4B5 +' + str(m_bonus[0]) + ' (вкрадено фокусником!)'
 
             if hach2 == 1:
